requirements.txt
ipykernel
python-dotenv
Chroma
chroma-hnswlib
chromadb
duckduckgo-search
gradio
langchain
langchain-chroma
langchain-community
langchain-core
langchain-text-splitters
langgraph
langgraph-checkpoint
langsmith
numpy
pandas
pydantic
pydantic_core
pypdf
pyprojroot
PyYAML
SQLAlchemy
langchain_groq
langchain_huggingface